Renames ToggleCategoryStatus to SetCategoryIsEnabled
