+++ src/Adapter/Category/CommandHandler/SetCategoryIsEnabledHandler.php
@@ -27,17 +27,15 @@
 namespace PrestaShop\PrestaShop\Adapter\Category\CommandHandler;
 
 use Category;
-use PrestaShop\PrestaShop\Core\Domain\Category\Command\ToggleCategoryStatusCommand;
-use PrestaShop\PrestaShop\Core\Domain\Category\CommandHandler\ToggleCategoryStatusHandlerInterface;
+use PrestaShop\PrestaShop\Core\Domain\Category\Command\SetCategoryIsEnabledCommand;
+use PrestaShop\PrestaShop\Core\Domain\Category\CommandHandler\SetCategoryIsEnabledHandlerInterface;
 use PrestaShop\PrestaShop\Core\Domain\Category\Exception\CannotUpdateCategoryStatusException;
 use PrestaShop\PrestaShop\Core\Domain\Category\Exception\CategoryNotFoundException;
 
 /**
- * Class ToggleCategoryStatusHandler.
- *
  * @internal
  */
-final class ToggleCategoryStatusHandler implements ToggleCategoryStatusHandlerInterface
+final class SetCategoryIsEnabledHandler implements SetCategoryIsEnabledHandlerInterface
 {
     /**
      * {@inheritdoc}
@@ -45,21 +43,22 @@ final class ToggleCategoryStatusHandler implements ToggleCategoryStatusHandlerIn
      * @throws CategoryNotFoundException
      * @throws CannotUpdateCategoryStatusException
      */
-    public function handle(ToggleCategoryStatusCommand $command)
+    public function handle(SetCategoryIsEnabledCommand $command)
     {
-        $entity = new Category($command->getCategoryId()->getValue());
+        $categoryId = $command->getCategoryId()->getValue();
+        $entity = new Category($categoryId);
 
         if (!$entity->id) {
             throw new CategoryNotFoundException(
                 $command->getCategoryId(),
-                sprintf('Category with id "%s" was not found', $command->getCategoryId()->getValue())
+                sprintf('Category with id "%s" was not found', $categoryId)
             );
         }
 
         if (!$entity->toggleStatus()) {
             throw new CannotUpdateCategoryStatusException(sprintf(
                 'Cannot update status for category with id "%s"',
-                $command->getCategoryId()->getValue()
+                $categoryId
             ));
         }
     }
